requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-multipart>=0.0.5
pandas>=1.3.0
google-cloud-bigquery>=2.34.0
google-auth>=2.0.0
